* Move matrix-matrix multiply to new cross-built project with breeze + natives impl for the JVM version
