+++ math/shared/src/main/scala/com/cibo/evilplot/numeric/MatrixOperationsI.scala
@@ -1,0 +1,7 @@
+package com.cibo.evilplot.numeric
+
+trait MatrixOperationsI {
+  def matrixMatrixTransposeMult(
+    a: Array[Array[Double]],
+    b: Array[Array[Double]]): Array[Array[Double]]
+}
